examples/pydantic-ai/requirements.txt
pydantic-ai>=0.0.9
sidclaw>=0.1.2
python-dotenv>=1.0.0
